implemented heuristic geneprob

+++ src/global.f90
@@ -63,7 +63,7 @@ module Global
     integer :: NRMmem, shell, shellmax, shellWarning
 
     type(AlphaPhaseResultsContainer) :: apResults
-    double precision, allocatable :: GenosProbs(:,:,:) !< output of geneprob
+    real(kind=real64), allocatable :: GenosProbs(:,:,:) !< output of geneprob
